# 3-pie-fire-dire/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Pie Fire Dire application.",
        "title": "Pie Fire Dire",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/{version}/beef": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Beef"
                ],
                "summary": "Get beef details",
                "parameters": [
                    {
                        "enum": [
                            "v1"
                        ],
                        "type": "string",
                        "description": "App Version",
                        "name": "version",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/beef.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/{version}/beef/summary": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Beef"
                ],
                "summary": "Get beef summary",
                "parameters": [
                    {
                        "enum": [
                            "v1"
                        ],
                        "type": "string",
                        "description": "App Version",
                        "name": "version",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/beef.GetSummaryResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/beef.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "beef.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "beef.GetSummaryResponse": {
            "type": "object",
            "properties": {
                "beef": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        }
    }
}